# Conflicts:
#	src/main/java/io/github/mzmine/datamodel/FeatureIdentity.java
#	src/main/java/io/github/mzmine/datamodel/features/types/annotations/ManualAnnotationType.java
#	src/main/java/io/github/mzmine/modules/dataprocessing/id_localcsvsearch/LocalCSVDatabaseSearchTask.java
#	src/main/java/io/github/mzmine/modules/visualization/spectra/simplespectra/spectraidentification/customdatabase/SpectraIdentificationCustomDatabaseTask.java

+++ src/main/java/io/github/mzmine/util/StreamCopy.java
@@ -85,7 +85,7 @@ public class StreamCopy {
       return 0;
     if (copiedLength >= totalLength)
       return 1.0;
-    return copiedLength / totalLength;
+    return copiedLength / (double)totalLength;
   }
 
   /**
